requirements.txt
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
pdf2image==1.16.3
easyocr==1.7.0
numpy==1.24.3
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
